Standard Update 21.02.2020

+++ Homeworks/day.02/day.02 homework - 08.py
@@ -1,7 +1,7 @@
 num = input('Enter your number')
 if int(num)== 0:
     print('You are not allowed to divide by 0')
-elif int(num)%3 == 0 or float(num)%5 == 0 or float(num)%7 == 0:
+elif int(num)%3 == 0 or int(num)%5 == 0 or int(num)%7 == 0:
     print('Your number can be divided by 3,5 or 7')
 else:
     print("Your number can't be divided by 3, 5 or 7")
